sipxAudioGetInputEnergy and sipxAudioGetOutputEnergy now return double instead of unsigned int. For VU mode, unsigned int was useless as the average value is very low. This will make VU mode more useful.

+++ sipXmediaLib/include/mp/MpAudioDriverBase.h
@@ -388,14 +388,14 @@ public:
     * Gets volume meter reading for input of given stream.
     */
    virtual OsStatus getInputVolumeMeterReading(MpAudioStreamId stream,
-                                               unsigned int& volume,
+                                               double& volume,
                                                MP_VOLUME_METER_TYPE type) const = 0;
 
    /**
    * Gets volume meter reading for output of given stream.
    */
    virtual OsStatus getOutputVolumeMeterReading(MpAudioStreamId stream,
-                                                unsigned int& volume,
+                                                double& volume,
                                                 MP_VOLUME_METER_TYPE type) const = 0;
 
    /**
